+++ frontend/package.json
@@ -9,9 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "12.3.1",
+    "bootstrap": "^5.2.1",
+    "next": "^12.3.1",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "react-modal": "^3.15.1"
   },
   "devDependencies": {
     "@types/node": "18.7.23",
@@ -20,7 +22,6 @@
     "eslint": "8.24.0",
     "eslint-config-next": "12.3.1",
     "postcss": "^8.4.16",
-    "tailwindcss": "^3.1.8",
     "typescript": "^4.8.4"
   }
 }
