+++ frontend/pages/_app.tsx
@@ -1,5 +1,8 @@
 import '../styles/globals.css'
 
+// add bootstrap css 
+import 'bootstrap/dist/css/bootstrap.css'
+
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />
 }
